rollback to the migration of the last triggers of the look_up table

+++ migrations/20250311054626-update-triggers-for-lookup-table.js
@@ -21,7 +21,9 @@ module.exports = {
       AFTER DELETE ON files
       FOR EACH ROW
       EXECUTE FUNCTION delete_file_from_look_up_table();
+    `);
 
+    await queryInterface.sequelize.query(`
       CREATE OR REPLACE FUNCTION delete_folder_from_look_up_table()
       RETURNS TRIGGER AS $$
       BEGIN
@@ -42,6 +44,7 @@ module.exports = {
     `);
   },
 
+  // Rollback to: migrations/20231004145835-add-triggers-for-lookup.js
   async down(queryInterface, Sequelize) {
     await queryInterface.sequelize.query(`
       CREATE OR REPLACE FUNCTION delete_file_from_look_up_table()
@@ -51,7 +54,8 @@ module.exports = {
         DELETE FROM look_up
         WHERE
           item_type = 'file' 
-          AND item_id = OLD.uuid;
+          AND item_id = OLD.uuid 
+          AND user_id = OLD.user_id;
     
         RETURN OLD;
       END;
@@ -71,7 +75,8 @@ module.exports = {
         DELETE FROM look_up
         WHERE
           item_type = 'folder' 
-          AND item_id = OLD.uuid;
+          AND item_id = OLD.uuid 
+          AND user_id = OLD.user_id;
     
         RETURN OLD;
       END;
